massive refactor / cleanup of typings.
