--- src/components/TableV2/index.ts
@@ -1,12 +1,0 @@
-export { default } from "./TableExample";
-export {
-  BaseTableWrapper,
-  BaseWrapper,
-  BaseTitle,
-  BaseEmptyRow,
-  BaseTableHeadRow,
-  BaseTableBody,
-  BaseTableRow,
-  BaseHeadCell,
-  BaseTableCell,
-} from "./Table.styles";
